JPA data type changes
Upgrade to hibernate 5.2.x to correctly serialize ZonedDateTime

+++ src/main/java/com/ianhattendorf/sensi/sensimonitor/domain/StatusRepository.java
@@ -2,5 +2,5 @@ package com.ianhattendorf.sensi.sensimonitor.domain;
 
 import org.springframework.data.repository.CrudRepository;
 
-public interface StatusRepository extends CrudRepository<Status, Long> {
+public interface StatusRepository extends CrudRepository<Status, Integer> {
 }
